+++ src/index.ts
@@ -5,6 +5,7 @@ import StButtonGroup from './components/button-group/index.vue';
 import StButton from './components/button/index.vue';
 import StCheckbox from './components/checkbox/index.vue';
 import StCol from './components/col/index.vue';
+import StCollapser from './components/collapser/index.vue';
 import StIcon from './components/icon/index.vue';
 import StInput from './components/input/index.vue';
 import { StNotificationService } from './components/notification/service';
@@ -14,7 +15,11 @@ import StPopper from './components/popper/index.vue';
 import StRadio from './components/radio/index.vue';
 import StRow from './components/row/index.vue';
 import StScrollbar from './components/scrollbar/index.vue';
+import StSelect from './components/select/index.vue';
 import StSwitch from './components/switch/index.vue';
+import StTable from './components/table/index.vue';
+import StTabsCollapsed from './components/tabs-collapsed/index.vue';
+import StTabs from './components/tabs/index.vue';
 import StTextarea from './components/textarea/index.vue';
 
 const install = (Vue: VueConstructor) => {
@@ -23,16 +28,21 @@ const install = (Vue: VueConstructor) => {
   Vue.component('st-button-group', StButtonGroup);
   Vue.component('st-checkbox', StCheckbox);
   Vue.component('st-col', StCol);
+  Vue.component('st-collapser', StCollapser);
   Vue.component('st-icon', StIcon);
   Vue.component('st-input', StInput);
   Vue.component('st-notifications-group', StNotificationsGroup);
   Vue.component('st-row', StRow);
   Vue.component('st-textarea', StTextarea);
   Vue.component('st-radio', StRadio);
-  Vue.component('st-scrollbar', StScrollbar);
-  Vue.component('st-switch', StSwitch);
   Vue.component('st-pagination', StPagination);
   Vue.component('st-popper', StPopper);
+  Vue.component('st-select', StSelect);
+  Vue.component('st-scrollbar', StScrollbar);
+  Vue.component('st-switch', StSwitch);
+  Vue.component('st-table', StTable);
+  Vue.component('st-tabs', StTabs);
+  Vue.component('st-tabs-collapsed', StTabsCollapsed);
 
   /* eslint-disable-next-line no-param-reassign */
   Vue.prototype.$stNotification = StNotificationService;
@@ -45,6 +55,7 @@ export default {
   StButtonGroup,
   StCheckbox,
   StCol,
+  StCollapser,
   StIcon,
   StInput,
   StNotificationsGroup,
@@ -52,4 +63,10 @@ export default {
   StTextarea,
   StPagination,
   StPopper,
+  StSelect,
+  StScrollbar,
+  StSwitch,
+  StTable,
+  StTabs,
+  StTabsCollapsed,
 };